requirements.txt
fastapi==0.100.0
uvicorn[standard]
sqlalchemy
pymysql
httpx
apscheduler
python-dotenv